+++ src/components/shipment/creation/index.styles.tsx
@@ -3,7 +3,6 @@ import styled from 'styled-components';
 export const FormWrapper = styled.div`
   display: flex;
   flex-direction: column;
-  align-items: center;
   margin: 0 auto;
   width: 800px;
   padding: 20px;
